prepare, patches: Remove use of scmversion and dirty string altogether
Command `scripts/setlocalversion --save-scmversion' was made to only execute
for <= 6.2-based kernel trees due to modifications done in 6.3.

Just remove it altogether along with a hack to not print -dirty string which
will always be embedded ever since linux-tkg switched to clone git repository,
because the tree will always be "dirty" the way we are doing.

Signed-off-by: Albert I <kras@raphielgang.org>